pyproject.toml
[tool.poetry]
name = "bayes-air"
version = "0.0.1"
description = "A framework for Bayesian modeling and analysis of air transportation network dynamics"
authors = ["Charles Dawson <charles.dwsn@gmail.com>"]
license = "BSD 3-Clause"
readme = "README.md"
packages = [{include = "bayes_air"}]

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
pyro-ppl = "^1.8.6"
pandas = "^2.1.0"
numpy = "^1.26.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
torch = {version = "^2.1.0+cpu", source = "pytorch"}
graphviz = "^0.20.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.1"
ipykernel = "^6.25.2"
pytest = "^7.4.3"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
